+++ src/main/java/website/eccentric/tome/Proxy.java
@@ -1,9 +1,0 @@
-package website.eccentric.tome;
-
-import net.minecraft.item.ItemStack;
-
-public class Proxy {
-
-    public void tomeScreen(ItemStack tome) { }
-
-}