+++ src/main/java/website/eccentric/tome/events/OpenTomeEvent.java
@@ -1,0 +1,12 @@
+package website.eccentric.tome.events;
+
+import net.minecraft.item.ItemStack;
+import net.minecraftforge.eventbus.api.Event;
+
+public class OpenTomeEvent extends Event {
+    public final ItemStack tome;
+
+    public OpenTomeEvent(ItemStack tome) {
+        this.tome = tome;
+    }
+}